APK provider to bundle tiles with app itself
With this change, tiles can be bundled into assets of apk itself. With
this people can easily make apps that load first couple of zoom levels
from apk and for the rest normal network/cache behavior be used.

tiles should be put into assets directory like this:

  Mapnik/11/1316/806.png

Note: First parameter is the name of tile source.