Update CacheManager.java - Add Alert-Confirmation before canceling
In my opinion the user should be asked, if he wants to cancel the map download. Since it may be the case that the cancel action was not intended.
I added an option "verifyCancel" which can be set to true or false. On true it will display an alert asking for confirmation.