feature/#977: Sometimes CacheManager#downloadAreaAsync too slow
For downloadAreaAsync(...) all the theoretical tile points were calculated in the
CacheManagerTask constructor. As for big bounding boxes the number of
points can be quite huge - it will take significant time to calculate all the points.
On top of that, the number of points can be so huge that the tiles list won't be
able to store all of them.

To fix this, create an iterable holding only 2 points for each zoom level - coordinates
of the upper left and lower right corners of the bounding box. Then, when iterating
calculate the point and return it.

At last, introduce unit tests for the new methods.