Feature #182: Fastest marker display ever ;)
The general idea was to get rid of most of the canvas manipulations:
* the canvas scale part became useless when we started using double zooms instead month ago (scale always equals to 1)
* the canvas rotation part can be prevented most of the time. Before we decided to display, we used to check if the rotated canvas (rotation using save, rotate, restore) matched the marker area. Now we check if the canvas matches the _rotated_ marker area (the rotated `Rect` is computed with ad hoc optimized code).

To my standards the performances are very good.
And I believe I also fixed a "click of the marker" bug in the process.

Demo "More Samples / Data / Itemized overlay at Scale" now has 5K markers (instead of 500) and rotation gesture, for performance test purpose.
One remark: the info window follows the map orientation, but the icon does not. What is the expected behavior?

Demo "OSMapView with ItemizedOverlay" now has rotation gesture, for display and click test purpose.

Potential error: we use class members in order to avoid recreating `Rect`s and `Point`s, and they are used for the marker display AND for the click detection as well. Do you guys think we should play it safe - I mean, we're only talking about a couple of `Rect`s and a `Point` to be created for each display?

New class:
* `RectLTest`: unit test class for `RectL`, more specifically on the new methods around `getBounds`

Impacted classes:
* `ItemizedIconOverlay`: refactored `activateSelectedItems` using new method `ItemizedOverlay.isEventOnItem`
* `ItemizedOverlay`: : removed `getMatrix` and related variables from `draw` as it was only used for scale values that are now always 1; removed `synchronized` from `boundToHotspot` by removing use of a class member; optimized method `onDrawItem` by not using canvas manipulation if not needed and using new member `Rect mOrientedMarkerRect`; added method `isEventOnItem` as refactoring; refactored `onSingleTapConfirmed` using new method `isEventOnItem`
* `MyLocationNewOverlay`: [slightly unrelated refactoring] removed `getMatrix` and related variables from `drawMyLocation` as it was only used for scale values that are now always 1
* `RectL`: new methods `getBounds` as a rough computation of the smaller RectL/Rect that contains a rotated RectL/Rect; also added helper methods `getRotatedX` and `getRotatedY`
* `SampleWithMinimapItemizedoverlay`: added the rotation gesture for test purposes
* `SampleWithMinimapItemizedOverlayWithScale`: added the rotation gesture and rose the number of markers from 500 to 5K for test purposes