feature/#977: Fix getTilesCoverageIterable method
The method was always skipping the last column of y because
if (x >= lowerRight.x - 1) was the only condition used for
checking whether to switch to another zoom or not. So, it's
fixed now and supported by the modification of the unit test.

Also, use Rect as a return value for getTilesRect.