bug/#989 - adding an increasing delay in tile fetch repetition
When for some reason we could not get an up-to-date version of a tile, the tile fetch process went in an endless loop in the search of this up-to-date version.
The endless loop is not good itself, and it's even worse when there's a provider that uses the network: tons of useless network tries.
The fix is to consider that if we can't get the up-to-date version, we wait before trying again. And the delay is increasing.

New classes:
* `Delay`: successive delays container
* `DelayTest`: the unit test class around `Delay`

Impacted classes:
* `MapTileDownloader`: unrelated optimization regarding the instanciation of `TileLoader`; considered `UnknownHostException` as similar to other `IOException`s
* `MapTileProviderArray`: added member `mTileDelays` as a `Map<tile, Delay>` and an exponential backoff delay array `mExponentialBackoffDurationInMillis`; added a check in `getMapTile` in order to delay the next tile fetch; adapted `mapTileRequest*` methods in order to affect the tile delay; created methods `shouldWait`, `nextDelay` and `removeDelay`