Bump support lib version to match compileSdk
24.2.0 deprecates support for API 8 and lower
26.0.0 raises minSDK to 14

So basically to be able to upgrade and comply with the playstore
rules minSDK MUST be raised to 14!

osmdroid-android doesn't have any dependencies to the support lib,
so we can leave the main minSDK at 8 and only set it to 14 for the
packages that do require support libraries (MapViewer, simple-map,
wms). This way the core lib will remain compatible down to API 8.
BUT this also means that running tests on AVDs below 14 will not be
possible and thus compatibility for API<14 will no longer be tested!