Bug/#1098 - Bump support lib to 26 (#1204)
* Bump support lib version to match compileSdk

24.2.0 deprecates support for API 8 and lower
26.0.0 raises minSDK to 14

So basically to be able to upgrade and comply with the playstore
rules minSDK MUST be raised to 14!

osmdroid-android doesn't have any dependencies to the support lib,
so we can leave the main minSDK at 8 and only set it to 14 for the
packages that do require support libraries (MapViewer, simple-map,
wms). This way the core lib will remain compatible down to API 8.
BUT this also means that running tests on AVDs below 14 will not be
possible and thus compatibility for API<14 will no longer be tested!

* Bump LeakCanary version to remove compiler warnings

* Manifest cleanup

values are overridden anyway by values set in build.gradle

* Mark license texts as non-translatable

to prevent lint warnings

* Fix travis build without signing config

In https://raw.githubusercontent.com/gradle-fury/gradle-fury/v1.1.4/gradle/android-support.gradle
we have a hack to prevent builds from failing due to missing signing
config. This is done by removing the signing task from the build if
no keystore/pass is provided.

At some point the signing task has been renamed from
"validateReleaseSigning" to "validateSigningRelease", thus obviously
breaking the build again.

This simply adds a copy of the relevant code from
android-support.gradle adapted to the new name.

This should be cleaned up in the future!

* Update Robolectric to fix build error