bug/#1217_2 - zoom and cache management refactoring
There used to be an unexpected removal of tiles from cache when zooming out. Not only did I remove this impact on cache, but I also made it less likely to happen again by downgrading cache method `remove` from `public` to `protected`.
The rest is about code optimization when zooming in and out.

Impacted classes:
* `MapTileCache`: downgraded method `remove` from `public` to `protected`, in order to avoid undesired cache management from outside
* `MapTileProviderBase`: edited the "is it worth?" check at the beginning of method `rescaleCache` in order to avoid eventually unnecessary code and instance creations; removed the removal of tile from cache in `ZoomOutTileLooper.computeTile`
* `MapView`: optimized the call to `MapTileProviderBase.rescaleCache` using the new `Rect mRescaleScreenRect` member
* `TilesOverlay`: replaced in method `protectDisplayedTilesForCache` the call to now removed inner class `CacheTileLooper` with a more direct code; created member `Rect mProtectedTiles` for optimization