bug/#1122 - removed confusion about long press in itemized overlay samples
Fixed the return value of `ItemizedIconOverlay.OnItemGestureListener.onItemLongPress` in the samples.
The default expected behavior is that if an item is pressed, the method should return `true`, in order to prevent other items to be potentially selected as well. This is already the case in method `onItemSingleTapUp`.
Of course, that doesn't prevent anyone to consider that `false` is the relevant return value, depending on the purpose of the app. But for the samples, `true` makes more sense.

For all the following classes, those are the changes:
* `onItemLongPress` now returns `true` when the first item is found
* minor refactoring in order to remove editor warnings

Impacted classes:
* `SampleMilitaryIconsItemizedIcons`
* `SampleWithMinimapItemizedOverlayWithFocus`
* `SampleWithMinimapItemizedOverlayWithScale`
* `SampleWithMinimapItemizedoverlay`