Bing map upgrade authentication and tile URL to https (#1436)
* Bing map authentication URL upgraded to https

The Bing maps authentication URL that is used for sending the application's Bing Maps Key to Microsoft was still using unencrypted HTTP, instead of HTTPS. 

I have verified that the URL via HTTPS works as expected, it was correctly returning a valid response using my personal Bing Maps key for the imagery-set "Road".

* Change the Bing Maps uriScheme to https

To get the tiles via https one have to set the uriScheme to `https`. 
By doing so the received `imageUrl` will contain an https url for downloading the map tiles. 
Therefore now the metadata request and the tile download will use a TLS protected communication channel.