feature/#1480 - new method PolyOverlayWithIW.setDowngradePixelSizes
If the poly is smaller (in pixels) than size1, draw a precomputed simplified list of segments.
If it's even smaller (than size2), draw even faster the poly as a rectangle.
Of course this is only optional; the default being to always display the regular poly.
The demo is included in `SampleShapeFile`, and the important methods here are `setDowngradeDisplay` and `setDowngradePixelSizes`.

New class:
* `ListPointAccepter`: a `PointAccepter` that builds a `List` of `PointL` as a list of long, long

Impacted classes:
* `LinearRing`: new members `mProjectedWidth` and `mProjectedHeight`; new method `computeDowngradePointList` that computes the list of points of a polyline that would be the projection of the GeoPoints on a centered size*size square; new method `resetPrecomputations`
* `PolyOverlayWithIW`: implemented the "fast draw of a list of segments" as less downgraded than a rectangle; new member `float[] mDowngradeSegments`; new method `isVisible(Paint)`; method `setDowngradePixelSizes` replaced `setDowngradeMaximumPixelSize`; minor refactoring
* `SampleShapeFile`: now using the new `PolyOverlayWithIW.setDowngradePixelSizes` method