feature/#1551 - stronger code
Thinking again about that, it could be very dangerous to change the color mapping on the fly. At least the way I did it.
When I recomputed I cleared the color array beforehand: if the polyline gets drawn while I'm recomputing, there may be crashes like "index not found".
The solution is to change the existing values instead: in the worst case, part of the colors are up-to-date and the second half hasn't been refreshed yet, but no crash. And with a call to `invalidate` after the container refresh the next canvas onDraw will be 100% up-to-date.

Impacted classes:
* `ColorMappingForScalarContainer`: method `refresh` now uses new method `ColorMappingForScalar.set(int, float)`
* `ColorMappingForScalar`: method `set(int, float)` replaces method `clear` that was dangerous