DelayTest.check: skip the test if a sleep call exceeds the expected wait time
The 'check' function's first call to Thread.sleep is intended to wait for _some_ but _not all_ of the current delay (backoff) duration.

In practice this seems to be causing some flaky unit test failures: my theory is that this occurs when the test thread is interrupted during the sleep, resulting in the DelayTest.sleep call sometimes _exceeding_ the delay duration.

To accommodate that possibility, we can use JUnit4's Assume[1] class to test whether the condition that Delay.shouldWait is true.

If the condition does _not_ hold -- that is, the DelayTest.sleep exceeded the expected delay duration -- then JUnit will skip the test (the condition for testing didn't hold, so we can't rely on the results).

[1] - https://junit.org/junit4/javadoc/4.12/org/junit/Assume.html